sample_vulnerable/requirements.txt
flask>=2.0
pyyaml>=5.4
requests>=2.28
django==2.2.14
lodash==4.17.20
pillow==8.0.0
